moved trakt cli stuff into its own file
